fix(app): use the Polish genitive in formatted dates
Self-review of the date helper this branch introduced. `month: 'short'`
rendered "8 maj 2026" in Polish, where the date genitive is "8 maja 2026".

CLDR's abbreviated Polish month names are genuine abbreviations for eleven
months — sty, lut, mar, kwi, cze, lip, sie, wrz, paź, lis, gru — and for May
the abbreviation *is* the whole nominative word, "maj". So the short form
looks perfect for eleven months of the year and reads wrong for the twelfth.
That is the worst shape a bug can have: invisible to anyone testing in
January, and it would have shipped across every date in the admin.

`month: 'long'` is unambiguously correct in both languages, and the places
these helpers are used — the setlist sidebar chip, the concert picker, the
setlist.fm result rows — are sparse labels with room for it.

Also records the English change explicitly rather than leaving it implicit:
the hand-rolled formatter these helpers replaced produced "08 May 2026", and
`en` through Intl gives "May 8, 2026". Delegating that choice to the platform
is the point of the change, but it is a visible difference and should be a
decision rather than a surprise.

Spec now pins the genitive for three months, May among them.

+++ app/src/utils/formatDate.ts
@@ -17,7 +17,18 @@ export function formatShortDate(date: string | null | undefined, locale: string)
   // An unparseable string is a data problem, not a formatting one — return it
   // untouched rather than rendering "Invalid Date" at the reader.
   if (Number.isNaN(parsed.getTime())) return date
-  return parsed.toLocaleDateString(locale, { day: 'numeric', month: 'short', year: 'numeric' })
+  // `month: 'long'`, not 'short'. Polish dates take the genitive — "8 maja",
+  // not "8 maj" — and CLDR's abbreviated form gives the nominative for May
+  // specifically, because the Polish abbreviation for that month *is* the
+  // whole word. Eleven months would have looked fine and the twelfth would
+  // have read wrong, which is the worst kind of bug to ship: invisible to
+  // anyone testing in January.
+  //
+  // This also changes the English rendering: the hand-rolled version these
+  // helpers replaced produced "08 May 2026", and `en` through Intl gives
+  // "May 8, 2026". That is what `en` means to the platform, and delegating
+  // the choice is the point of the change.
+  return parsed.toLocaleDateString(locale, { day: 'numeric', month: 'long', year: 'numeric' })
 }
 
 /** `245` → `4:05`. Seconds are data; the colon is not copy in any locale. */
